dbeaver/pro#7734 spec support
